+++ CHANGELOG.md
@@ -10,8 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add table and pretty_table log formatters
 - Add method and proc style validations
 - Refactor parameter modules and classes for more robust usages
+- Update `callback` and `middleware` to be module based
 
 ## [Unreleased]
+
+### Changed
+- Moved `Task::CALLBACKS` constant to `CallbackRegistry::TYPES`
 
 ## [1.0.1] - 2025-07-07